inital commit
creating a 2nd project, this time you are the room master, trying to
defeat the human as fast as possible.

got a room design with saws that bounce around when you click. worked on
timer at top too because i was getting some stuttering, turns out the
trace was causing it.

need to run the game occasionally on release to see if those are
development build performance issues.